# docs/swagger.json
{"info": {"version": "1.0", "title": "API"}, "security": [{"apikey": []}], "basePath": "/", "consumes": ["application/json"], "produces": ["application/json"], "responses": {"MaskError": {"description": "When any error occurs on mask"}, "ParseError": {"description": "When a mask can't be parsed"}}, "paths": {"/projects/": {"post": {"tags": ["default"], "responses": {"200": {"schema": {"$ref": "#/definitions/Project"}, "description": "Success"}}, "parameters": [{"name": "payload", "in": "body", "schema": {"$ref": "#/definitions/Project"}, "required": true}, {"name": "X-Fields", "in": "header", "type": "string", "description": "An optional fields mask", "format": "mask"}], "operationId": "post_projects"}, "get": {"tags": ["default"], "responses": {"200": {"schema": {"type": "array", "items": {"$ref": "#/definitions/Project"}}, "description": "Success"}}, "parameters": [{"name": "X-Fields", "in": "header", "type": "string", "description": "An optional fields mask", "format": "mask"}], "operationId": "get_projects"}}, "/projects/{name}/": {"parameters": [{"name": "name", "in": "path", "type": "string", "required": true}], "get": {"tags": ["default"], "responses": {"200": {"schema": {"$ref": "#/definitions/Project"}, "description": "Success"}}, "parameters": [{"name": "X-Fields", "in": "header", "type": "string", "description": "An optional fields mask", "format": "mask"}], "operationId": "get_project"}, "put": {"tags": ["default"], "responses": {"200": {"schema": {"$ref": "#/definitions/Project"}, "description": "Success"}}, "parameters": [{"name": "payload", "in": "body", "schema": {"$ref": "#/definitions/Project"}, "required": true}, {"name": "X-Fields", "in": "header", "type": "string", "description": "An optional fields mask", "format": "mask"}], "operationId": "put_project"}}, "/projects/{name}/_restart": {"post": {"tags": ["default"], "responses": {"200": {"description": "Success"}}, "operationId": "post_restart"}, "parameters": [{"name": "name", "in": "path", "type": "string", "required": true}]}, "/projects/{name}/_start": {"post": {"tags": ["default"], "responses": {"200": {"description": "Success"}}, "operationId": "post_start"}, "parameters": [{"name": "name", "in": "path", "type": "string", "required": true}]}, "/projects/{name}/_stop": {"post": {"tags": ["default"], "responses": {"200": {"description": "Success"}}, "operationId": "post_stop"}, "parameters": [{"name": "name", "in": "path", "type": "string", "required": true}]}}, "securityDefinitions": {"apikey": {"in": "header", "name": "X-API-KEY", "type": "apiKey"}}, "tags": [{"name": "default", "description": "Default namespace"}], "definitions": {"EnvKey": {"properties": {"key": {"type": "string"}, "value": {"type": "string"}}, "type": "object"}, "Project": {"properties": {"name": {"type": "string"}, "environment": {"type": "array", "items": {"$ref": "#/definitions/EnvKey"}}}, "type": "object"}}, "swagger": "2.0"}